+++ lib/view/for_app_setting/_dialog.dart
@@ -12,6 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+export './app_language_changer.dart'
+    show
+        AppLanguageChangerDialog,
+        AppLanguageChangerDialogResult,
+        showAppLanguageChangerDialog;
 export './app_setting_clear_cache.dart'
     show AppSettingClearCacheDialog, showAppSettingClearCacheDialog;
 export './app_setting_confirm_clear_db.dart';